refactor: make variant system generic and add Manage popup
Split pageVariants.ts into:
- variants.ts: generic Variant<T>, VariantStore<T>, useVariants<T>() hook
- singleCatVariants.ts: SingleCatSettings, parsing, migration, comparison

Move VariantBar to components/common/ with generic props (apiPath,
parsePayload). Replace inline Export/Import/Delete with a Manage popup
featuring export-with-copy, import, and two-click delete confirmation.

Add structural validation in loadVariantStore, clamp speedMultiplier
(0.1-10), deduplicate extendedModes, truncate name strings, and defer
old localStorage key cleanup until after persist succeeds.

Powered by human calories and mass GPU cycles.
